refactor: single source of truth for config
